+++ src/main/java/insane96mcp/survivalreimagined/module/farming/feature/BoneMeal.java
@@ -7,7 +7,7 @@ import insane96mcp.insanelib.base.config.Config;
 import insane96mcp.insanelib.base.config.LoadFeature;
 import insane96mcp.survivalreimagined.SurvivalReimagined;
 import insane96mcp.survivalreimagined.module.Modules;
-import insane96mcp.survivalreimagined.module.farming.block.CompostedFarmlandBlock;
+import insane96mcp.survivalreimagined.module.farming.block.RichFarmlandBlock;
 import insane96mcp.survivalreimagined.setup.SRBlocks;
 import insane96mcp.survivalreimagined.setup.SRItems;
 import insane96mcp.survivalreimagined.utils.Utils;
@@ -29,12 +29,12 @@ import net.minecraftforge.registries.RegistryObject;
 
 import java.util.Collections;
 
-@Label(name = "Bone meal", description = "Bone meal is no longer so OP and also Composted Farmland")
+@Label(name = "Bone meal", description = "Bone meal is no longer so OP and also Rich Farmland")
 @LoadFeature(module = Modules.Ids.FARMING)
 public class BoneMeal extends Feature {
 
-	public static final RegistryObject<CompostedFarmlandBlock> COMPOSTED_FARMLAND = SRBlocks.REGISTRY.register("composted_farmland", () -> new CompostedFarmlandBlock(BlockBehaviour.Properties.of(Material.DIRT).randomTicks().strength(0.6F).sound(SoundType.GRAVEL).isViewBlocking((state, blockGetter, pos) -> true).isSuffocating((state, blockGetter, pos) -> true)));
-	public static final RegistryObject<BlockItem> COMPOSTED_FARMLAND_ITEM = SRItems.REGISTRY.register("composted_farmland", () -> new BlockItem(COMPOSTED_FARMLAND.get(), new Item.Properties()));
+	public static final RegistryObject<RichFarmlandBlock> RICH_FARMLAND = SRBlocks.REGISTRY.register("rich_farmland", () -> new RichFarmlandBlock(BlockBehaviour.Properties.of(Material.DIRT).randomTicks().strength(0.6F).sound(SoundType.GRAVEL).isViewBlocking((state, blockGetter, pos) -> true).isSuffocating((state, blockGetter, pos) -> true)));
+	public static final RegistryObject<BlockItem> RICH_FARMLAND_ITEM = SRItems.REGISTRY.register("rich_farmland", () -> new BlockItem(RICH_FARMLAND.get(), new Item.Properties()));
 
 	private static final ResourceLocation BLACKLIST = new ResourceLocation(SurvivalReimagined.MOD_ID, "nerfed_bone_meal_blacklist");
 	@Config
@@ -43,16 +43,16 @@ public class BoneMeal extends Feature {
 	@Config(min = 0d, max = 1d)
 	@Label(name = "Bone Meal Fail Chance", description = "Makes Bone Meal have a chance to fail to grow crops. 0 to disable, 1 to disable Bone Meal.")
 	public static Double boneMealFailChance = 0d;
-	@Config
-	@Label(name = "Transform Farmland in Composted", description = "Bone meal used on Farmland transforms it into Composted Farmland.")
-	public static Boolean farmlandToComposted = true;
 
+	@Config
+	@Label(name = "Transform Farmland in Rich Farmland", description = "Bone meal used on Farmland transforms it into Rich Farmland.")
+	public static Boolean farmlandToRich = true;
 	@Config(min = 1)
-	@Label(name = "Composted Farmland Extra Ticks", description = "How many extra random ticks does Composted Farmland give to the crop sitting on top?")
-	public static Integer compostedFarmlandExtraTicks = 3;
+	@Label(name = "Rich Farmland Extra Ticks", description = "How many extra random ticks does Rich Farmland give to the crop sitting on top?")
+	public static Integer richFarmlandExtraTicks = 3;
 	@Config(min = 0d, max = 1d)
-	@Label(name = "Composted Farmland Chance to Decay", description = "Chance for a Composted farmland to decay back to farmland")
-	public static Double compostedFarmlandChanceToDecay = 0.05d;
+	@Label(name = "Rich Farmland Chance to Decay", description = "Chance for a Rich farmland to decay back to farmland")
+	public static Double richFarmlandChanceToDecay = 0.05d;
 
 	public BoneMeal(Module module, boolean enabledByDefault, boolean canBeDisabled) {
 		super(module, enabledByDefault, canBeDisabled);
@@ -68,8 +68,8 @@ public class BoneMeal extends Feature {
 				|| !this.isEnabled()
 				|| event.getLevel().isClientSide)
 			return;
-		if (farmlandToComposted && event.getBlock().is(Blocks.FARMLAND)){
-			event.getLevel().setBlockAndUpdate(event.getPos(), COMPOSTED_FARMLAND.get().defaultBlockState().setValue(FarmBlock.MOISTURE, event.getBlock().getValue(FarmBlock.MOISTURE)));
+		if (farmlandToRich && event.getBlock().is(Blocks.FARMLAND)){
+			event.getLevel().setBlockAndUpdate(event.getPos(), RICH_FARMLAND.get().defaultBlockState().setValue(FarmBlock.MOISTURE, event.getBlock().getValue(FarmBlock.MOISTURE)));
 			event.getEntity().swing(event.getEntity().getUsedItemHand(), true);
 			event.setResult(Event.Result.ALLOW);
 		}